Extract JDBC write-related configuration to separate class
