+++ plugin/trino-base-jdbc/src/main/java/io/trino/plugin/jdbc/JdbcMetadataSessionProperties.java
@@ -24,7 +24,6 @@ import javax.inject.Inject;
 import java.util.List;
 import java.util.Optional;
 
-import static io.trino.plugin.jdbc.JdbcMetadataConfig.MAX_ALLOWED_WRITE_BATCH_SIZE;
 import static io.trino.spi.StandardErrorCode.INVALID_SESSION_PROPERTY;
 import static io.trino.spi.session.PropertyMetadata.booleanProperty;
 import static io.trino.spi.session.PropertyMetadata.integerProperty;
@@ -37,8 +36,6 @@ public class JdbcMetadataSessionProperties
     public static final String AGGREGATION_PUSHDOWN_ENABLED = "aggregation_pushdown_enabled";
     public static final String TOPN_PUSHDOWN_ENABLED = "topn_pushdown_enabled";
     public static final String DOMAIN_COMPACTION_THRESHOLD = "domain_compaction_threshold";
-    public static final String WRITE_BATCH_SIZE = "write_batch_size";
-    public static final String NON_TRANSACTIONAL_INSERT = "non_transactional_insert";
 
     private final List<PropertyMetadata<?>> properties;
 
@@ -67,17 +64,6 @@ public class JdbcMetadataSessionProperties
                         TOPN_PUSHDOWN_ENABLED,
                         "Enable TopN pushdown",
                         jdbcMetadataConfig.isTopNPushdownEnabled(),
-                        false))
-                .add(integerProperty(
-                        WRITE_BATCH_SIZE,
-                        "Maximum number of rows to write in a single batch",
-                        jdbcMetadataConfig.getWriteBatchSize(),
-                        JdbcMetadataSessionProperties::validateWriteBatchSize,
-                        false))
-                .add(booleanProperty(
-                        NON_TRANSACTIONAL_INSERT,
-                        "Do not use temporary table on insert to table",
-                        jdbcMetadataConfig.isNonTransactionalInsert(),
                         false))
                 .build();
     }
@@ -108,16 +94,6 @@ public class JdbcMetadataSessionProperties
         return session.getProperty(DOMAIN_COMPACTION_THRESHOLD, Integer.class);
     }
 
-    public static int getWriteBatchSize(ConnectorSession session)
-    {
-        return session.getProperty(WRITE_BATCH_SIZE, Integer.class);
-    }
-
-    public static boolean isNonTransactionalInsert(ConnectorSession session)
-    {
-        return session.getProperty(NON_TRANSACTIONAL_INSERT, Boolean.class);
-    }
-
     private static void validateDomainCompactionThreshold(int domainCompactionThreshold, Optional<Integer> maxDomainCompactionThreshold)
     {
         if (domainCompactionThreshold < 1) {
@@ -129,15 +105,5 @@ public class JdbcMetadataSessionProperties
                 throw new TrinoException(INVALID_SESSION_PROPERTY, format("Domain compaction threshold (%s) cannot exceed %s", domainCompactionThreshold, max));
             }
         });
-    }
-
-    private static void validateWriteBatchSize(int maxBatchSize)
-    {
-        if (maxBatchSize < 1) {
-            throw new TrinoException(INVALID_SESSION_PROPERTY, format("%s must be greater than 0: %s", WRITE_BATCH_SIZE, maxBatchSize));
-        }
-        if (maxBatchSize > MAX_ALLOWED_WRITE_BATCH_SIZE) {
-            throw new TrinoException(INVALID_SESSION_PROPERTY, format("%s cannot exceed %s: %s", WRITE_BATCH_SIZE, MAX_ALLOWED_WRITE_BATCH_SIZE, maxBatchSize));
-        }
     }
 }
