+++ plugin/trino-base-jdbc/src/test/java/io/trino/plugin/jdbc/TestJdbcMetadataConfig.java
@@ -14,7 +14,6 @@
 package io.trino.plugin.jdbc;
 
 import com.google.common.collect.ImmutableMap;
-import io.airlift.configuration.ConfigurationFactory;
 import org.testng.annotations.Test;
 
 import java.util.Map;
@@ -22,8 +21,6 @@ import java.util.Map;
 import static io.airlift.configuration.testing.ConfigAssertions.assertFullMapping;
 import static io.airlift.configuration.testing.ConfigAssertions.assertRecordedDefaults;
 import static io.airlift.configuration.testing.ConfigAssertions.recordDefaults;
-import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class TestJdbcMetadataConfig
 {
@@ -35,9 +32,7 @@ public class TestJdbcMetadataConfig
                 .setJoinPushdownEnabled(false)
                 .setAggregationPushdownEnabled(true)
                 .setTopNPushdownEnabled(true)
-                .setDomainCompactionThreshold(32)
-                .setWriteBatchSize(1000)
-                .setNonTransactionalInsert(false));
+                .setDomainCompactionThreshold(32));
     }
 
     @Test
@@ -49,8 +44,6 @@ public class TestJdbcMetadataConfig
                 .put("aggregation-pushdown.enabled", "false")
                 .put("domain-compaction-threshold", "42")
                 .put("topn-pushdown.enabled", "false")
-                .put("write.batch-size", "24")
-                .put("insert.non-transactional-insert.enabled", "true")
                 .build();
 
         JdbcMetadataConfig expected = new JdbcMetadataConfig()
@@ -58,31 +51,8 @@ public class TestJdbcMetadataConfig
                 .setJoinPushdownEnabled(true)
                 .setAggregationPushdownEnabled(false)
                 .setTopNPushdownEnabled(false)
-                .setDomainCompactionThreshold(42)
-                .setWriteBatchSize(24)
-                .setNonTransactionalInsert(true);
+                .setDomainCompactionThreshold(42);
 
         assertFullMapping(properties, expected);
-    }
-
-    @Test
-    public void testWriteBatchSizeValidation()
-    {
-        assertThatThrownBy(() -> makeConfig(ImmutableMap.of("write.batch-size", "-42")))
-                .hasMessageContaining("write.batch-size: must be greater than or equal to 1");
-
-        assertThatThrownBy(() -> makeConfig(ImmutableMap.of("write.batch-size", "0")))
-                .hasMessageContaining("write.batch-size: must be greater than or equal to 1");
-
-        assertThatCode(() -> makeConfig(ImmutableMap.of("write.batch-size", "1")))
-                .doesNotThrowAnyException();
-
-        assertThatCode(() -> makeConfig(ImmutableMap.of("write.batch-size", "42")))
-                .doesNotThrowAnyException();
-    }
-
-    private static JdbcMetadataConfig makeConfig(Map<String, String> props)
-    {
-        return new ConfigurationFactory(props).build(JdbcMetadataConfig.class);
     }
 }
